Updated Oanda order ids from int to long

+++ Brokerages/Oanda/DataType/Transaction.cs
@@ -26,7 +26,7 @@ namespace QuantConnect.Brokerages.Oanda.DataType
     /// </summary>
     public class Transaction : Response
     {
-        public int id { get; set; }
+        public long id { get; set; }
         public int accountId { get; set; }
 		public string time { get; set; }
 		public string type { get; set; }
@@ -42,8 +42,8 @@ namespace QuantConnect.Brokerages.Oanda.DataType
 		public double pl { get; set; }
 		public double interest { get; set; }
 		public double accountBalance { get; set; }
-		public int tradeId { get; set; }
-	    public int orderId { get; set; }
+		public long tradeId { get; set; }
+	    public long orderId { get; set; }
 		public TradeData tradeOpened { get; set; }
 		public TradeData tradeReduced { get; set; }
 		public string reason { get; set; }
